Accommodate changed root topic
In 2.19 the root topic was changed from /msh/2 to /msh/<region>/2. Now we need to subscribe to _all_ packets to get the location specific ones.

This means one of the first steps is to check if it's a raw ProtoBuf packet (typically with a topic matching '.*/c/.*') and ignore it if not.